More SEGFAULT safety

+++ gst/gst_device_monitor.go
@@ -62,7 +62,7 @@ func (d *DeviceMonitor) GetDevices() []*Device {
 	if glist == nil {
 		return nil
 	}
-	goList := glib.WrapList(uintptr(unsafe.Pointer(glist)))
+	goList := glib.WrapList(unsafe.Pointer(glist))
 	out := make([]*Device, 0)
 	goList.Foreach(func(item interface{}) {
 		pt := item.(unsafe.Pointer)
