+++ gst/gst_element.go
@@ -333,7 +333,7 @@ func (e *Element) GetPadTemplates() []*PadTemplate {
 	if glist == nil {
 		return nil
 	}
-	goList := glib.WrapList(uintptr(unsafe.Pointer(glist)))
+	goList := glib.WrapList(unsafe.Pointer(glist))
 	out := make([]*PadTemplate, 0)
 	goList.Foreach(func(item interface{}) {
 		pt := item.(unsafe.Pointer)
